@@ -4,5 +4,5 @@ namespace Ipag\Http;
 
 interface OnlyPostHttpClientInterface
 {
-    public function __invoke($url, array $headers = array(), array $fields = array());
+    public function __invoke($url, array $headers = [], array $fields = []);
 }
