@@ -18,11 +18,11 @@ final class ConsultSerializer implements Serializer
 
     public function serialize()
     {
-        return array(
+        return [
             'identificacao' => urlencode($this->transaction->getIpag()->getAuthentication()->getIdentification()),
-            'transId' => urlencode($this->transaction->getTid()),
-            'url_retorno' => urlencode($this->transaction->getOrder()->getCallbackUrl()),
-            'retorno_tipo' => urlencode('xml'),
-        );
+            'transId'       => urlencode($this->transaction->getTid()),
+            'url_retorno'   => urlencode($this->transaction->getOrder()->getCallbackUrl()),
+            'retorno_tipo'  => urlencode('xml'),
+        ];
     }
 }
