@@ -22,27 +22,27 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint(Endpoint::SANDBOX);
 
-        $this->assertEquals($endpoint->payment(), Endpoint::SANDBOX . Endpoint::PAYMENT);
+        $this->assertEquals($endpoint->payment(), Endpoint::SANDBOX.Endpoint::PAYMENT);
     }
 
     public function testGetConsultEndpointUrl()
     {
         $endpoint = new Endpoint(Endpoint::SANDBOX);
 
-        $this->assertEquals($endpoint->consult(), Endpoint::SANDBOX . Endpoint::CONSULT);
+        $this->assertEquals($endpoint->consult(), Endpoint::SANDBOX.Endpoint::CONSULT);
     }
 
     public function testGetCancelEndpointUrl()
     {
         $endpoint = new Endpoint(Endpoint::SANDBOX);
 
-        $this->assertEquals($endpoint->cancel(), Endpoint::SANDBOX . Endpoint::CANCEL);
+        $this->assertEquals($endpoint->cancel(), Endpoint::SANDBOX.Endpoint::CANCEL);
     }
 
     public function testGetCaptureEndpointUrl()
     {
         $endpoint = new Endpoint(Endpoint::SANDBOX);
 
-        $this->assertEquals($endpoint->CAPTURE(), Endpoint::SANDBOX . Endpoint::CAPTURE);
+        $this->assertEquals($endpoint->CAPTURE(), Endpoint::SANDBOX.Endpoint::CAPTURE);
     }
 }
